backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
aiofiles<1.0.0
aiohttp==3.9.3
python-dotenv==1.0.1
asyncpraw==7.7.1
openai==1.12.0
httpx==0.26.0
beautifulsoup4==4.12.3
jinja2==3.1.2
markupsafe==2.1.5
aiosqlite>=0.17.0
